requirements.txt
python-decouple==3.1
python-telegram-bot==11.1.0
requests==2.16.0
